requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
tinydb==4.8.0
google-cloud-speech
gunicorn==21.2.0